mnt: extract flight data exporters into FlightDataExporter with deprecation wrappers
