+++ rocketpy/simulation/__init__.py
@@ -2,3 +2,4 @@ from .flight import Flight
 from .flight_data_importer import FlightDataImporter
 from .monte_carlo import MonteCarlo
 from .multivariate_rejection_sampler import MultivariateRejectionSampler
+from .flight_data_exporter import FlightDataExporter
